+++ src/tests.rs
@@ -64,63 +64,3 @@ fn test_trail() {
     assert_eq!(trail.is_empty(), true);
     assert_eq!(trail.conflict_literals.borrow().len(), 0);
 }
-
-#[test]
-fn test_encoding_rules_shouldbe_ok() {
-    // Doesn't encompass all cases
-    use crate::cnf::cnf_encoding_rules_ok;
-
-    let cell_at_least_one = true;
-    let cell_at_most_one = false;
-    let sudoku_has_all_values = false;
-    let sudoku_has_unique_values = true;
-
-    assert!(cnf_encoding_rules_ok(
-        cell_at_least_one,
-        cell_at_most_one,
-        sudoku_has_all_values,
-        sudoku_has_unique_values
-    ));
-
-    let cell_at_least_one = false;
-    let cell_at_most_one = true;
-    let sudoku_has_all_values = true;
-    let sudoku_has_unique_values = false;
-
-    assert!(cnf_encoding_rules_ok(
-        cell_at_least_one,
-        cell_at_most_one,
-        sudoku_has_all_values,
-        sudoku_has_unique_values
-    ));
-}
-
-#[test]
-fn test_encoding_rules_shouldbe_not_ok() {
-    // Doesn't encompass all cases
-    use crate::cnf::cnf_encoding_rules_ok;
-
-    let cell_at_least_one = true;
-    let cell_at_most_one = true;
-    let sudoku_has_all_values = false;
-    let sudoku_has_unique_values = false;
-
-    assert!(!cnf_encoding_rules_ok(
-        cell_at_least_one,
-        cell_at_most_one,
-        sudoku_has_all_values,
-        sudoku_has_unique_values
-    ));
-
-    let cell_at_least_one = true;
-    let cell_at_most_one = false;
-    let sudoku_has_all_values = false;
-    let sudoku_has_unique_values = false;
-
-    assert!(!cnf_encoding_rules_ok(
-        cell_at_least_one,
-        cell_at_most_one,
-        sudoku_has_all_values,
-        sudoku_has_unique_values
-    ));
-}
